done
fixed the header, all properties are in the top of the script.

reordered the page (the menu isn't a part of the header).

changed the left/right icons to ones from font-awesome, is it possible
to change it's color?

added a gradient in the menu, not sure about it, what do you think?

added the dark gray (#242424) to the list.

need to do the smart menu that stays always on top.

need to add redirect to the menu things.

need to set the logo size according to the web size. (I will try to do
it)

need to add loading screen. (i will do this too, if you've got a good
library of loading animations is will be nice)

the web is developing very well, good job!!!