music player (read the problem)
I don't like this widget but this is the only one working fine.
for now i use yahav's music because we don't have nothing on soundcloud
yet.
There is the minimal skin widget in SoundCloud that is prefect for us
but I can't find it in their website. They might removed it or
something.
I have more other players but they aren't what I'm searching for. If you
know how to use different skins on the SoundCloud API it will be great.